+++ CHANGELOG.md
@@ -1,12 +1,13 @@
-### V0.1.13
+### V1.0.0
 * add ngcanvasgauges and new modern theme (thanks godind!)
 * new Theme change event allowing widgets to subscribe to theme changes (thanks godind!)
 * Historical Graph color change on theme update
 * Update WidgetString to use canvas like numeric widget for dynamic font size
 * Latitude/Longitude Unit format
 * Load / Save config from server!
-* Notifications service
+* Notifications service for application status (snackbar)
 * Deleting the last widget in a page deletes the page!
+* Signalk Notifications in menubar!
 
 ### V0.1.12
 * iFrame to embed something in kip